requirements.txt
stable-baselines3[extra]>=2.0.0
gymnasium>=0.29.0
selenium>=4.15.0
opencv-python>=4.8.0
numpy>=1.24.0
pillow>=10.0.0
webdriver-manager>=4.0.0
requests>=2.28.0
beautifulsoup4>=4.11.0
